build: make -Werror fatal on macOS again, as it already is on Linux
The if(APPLE) branch that stripped -Werror was a bring-up measure with its own
TODO to re-tighten. Its cost was concrete: a warning that FAILS the Linux build
is merely printed on macOS, so a clean-looking local build can still break CI.
That happened, on the commit before this one. Removing the branch removes the
whole class rather than managing it.

All 482 PJ-governed translation units are now warning-clean under Apple Clang.
The work split three ways.

Vendored headers came first, because they set the ceiling: -Wconversion firing
inside qwt_series_data.h or nanocdr.hpp is attributed to whichever PJ file
included it, and no cast in OUR code can silence a warning in THEIRS. Both are
now marked SYSTEM from the root CMakeLists rather than by editing the vendored
trees — this is our policy about code we do not own, and keeping those trees
pristine keeps re-syncing them a copy rather than a merge. That one change also
retired the global -Wno-invalid-constexpr: it existed solely because nanocdr's
getCurrentEndianness() type-puns a union, and SYSTEM covers it, so a
project-wide suppression of a real diagnostic is gone.

Most of the rest were half-done casts, the same pattern as the previous commit.

Three were not cosmetic. derived_engine.cpp's findFirstLeaf is static and
referenced only by ITSELF — nthLeafPrimitive directly below it is what the
engine calls — so Clang's "not needed and will not be emitted" was correct and
it is deleted rather than suppressed. media_viewer_widget.cpp cast to int
before passing to QRhi APIs that take quint32 (newBuffer, setSize,
updateDynamicBuffer), a cast that actively misstated the parameter type; it now
casts to quint32. And a test lambda captured a constant expression it never
needed to capture.

check_werror.py replays compile_commands.json entries with -Werror appended and
the object discarded, for checking a subset without a full build. It only
checks translation units that actually carry PJ_WARNING_FLAGS, detected via
-Wold-style-cast: the flags are per-target, so vendored targets building their
own sources never receive -Werror and reporting them was a false alarm. It is
Apple-Clang-accurate, NOT GCC-accurate — it catches the shared warning set,
where the cross-platform breaks have come from, but a GCC-only diagnostic still
surfaces only on Linux.

+++ pj_app/src/MainWindow.cpp
@@ -3675,7 +3675,8 @@ QDomElement MainWindow::saveRightPanelState(QDomDocument& doc) const {
   if (width_button_group_ != nullptr) {
     const int id = width_button_group_->checkedId();
     if (id >= 0 && id < static_cast<int>(kWidthButtonSpecs.size())) {
-      element.setAttribute(QStringLiteral("width"), QString::number(kWidthButtonSpecs[id].second, 'g'));
+      element.setAttribute(
+          QStringLiteral("width"), QString::number(kWidthButtonSpecs[static_cast<std::size_t>(id)].second, 'g'));
     }
   }
 
@@ -3798,7 +3799,7 @@ void MainWindow::restoreRightPanelState(const QDomElement& element) {
     const double wanted = element.attribute(QStringLiteral("width")).toDouble(&ok);
     if (ok) {
       for (int i = 0; i < static_cast<int>(kWidthButtonSpecs.size()); ++i) {
-        if (qFuzzyCompare(kWidthButtonSpecs[i].second, wanted)) {
+        if (qFuzzyCompare(kWidthButtonSpecs[static_cast<std::size_t>(i)].second, wanted)) {
           checkGroupButton(width_button_group_, i);
           break;
         }
@@ -4476,8 +4477,8 @@ void MainWindow::buildLocalToolbar() {
   width_button_group_->setExclusive(true);
   const int initial_width_id = 0;
   for (int i = 0; i < static_cast<int>(kWidthButtonSpecs.size()); ++i) {
-    auto* btn =
-        curve_width_header_->parentWidget()->findChild<QToolButton*>(QString::fromLatin1(kWidthButtonSpecs[i].first));
+    auto* btn = curve_width_header_->parentWidget()->findChild<QToolButton*>(
+        QString::fromLatin1(kWidthButtonSpecs[static_cast<std::size_t>(i)].first));
     if (btn == nullptr) {
       continue;
     }
@@ -4585,7 +4586,7 @@ void MainWindow::applyActivePlotWidth(double width) {
   // becomes the width any curve added later inherits. Map the toolbar's pixel
   // value to the LineWidth enum it mirrors (kWidthButtonSpecs index == enum).
   for (int i = 0; i < static_cast<int>(kWidthButtonSpecs.size()); ++i) {
-    if (qFuzzyCompare(kWidthButtonSpecs[i].second, width)) {
+    if (qFuzzyCompare(kWidthButtonSpecs[static_cast<std::size_t>(i)].second, width)) {
       plot->setLineWidth(static_cast<LineWidth>(i));
       onUndoableChange();
       break;
